Main-website/backend/requirements.txt
absl-py==2.2.2
addict==2.4.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
basicsr==1.4.2
certifi==2022.12.7
cffi==1.17.1
charset-normalizer==2.1.1
click==8.1.8
contourpy==1.3.0
cycler==0.12.1
dnspython==2.7.0
exceptiongroup==1.2.2
facexlib==0.3.0
fastapi==0.115.12
filelock==3.18.0
filterpy==1.4.5
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
future==1.0.0
gfpgan==1.3.8
grpcio==1.71.0
h11==0.16.0
huggingface==0.0.1
huggingface-hub==0.30.2
idna==3.4
imageio==2.37.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
jax==0.4.30
jaxlib==0.4.30
Jinja2==3.1.6
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
lmdb==1.6.2
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.9.4
# mediapipe==0.10.21
mediapipe==0.10.9
ml_dtypes==0.5.1
motor==3.7.0
mpmath==1.3.0
networkx==3.2.1
numba==0.60.0
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pika==1.3.2
pillow==11.2.1
platformdirs==4.3.7
# protobuf==4.25.7
protobuf>=3.20,<4.0
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pymongo==4.12.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
PyYAML==6.0.2
realesrgan==0.3.0
requests==2.28.1
scikit-image==0.24.0
scipy==1.13.1
sentencepiece==0.2.0
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.1
starlette==0.46.2
sympy==1.13.3
tb-nightly==2.20.0a20250424
tensorboard-data-server==0.7.2
tifffile==2024.8.30
tomli==2.2.1
torch==2.7.0
torchaudio==0.11.0
torchvision==0.12.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==1.26.13
uvicorn==0.34.2
Werkzeug==3.1.3
yapf==0.43.0
zipp==3.21.0
